moved the clocking initialization to the PRSM module. Doesn't compile yet.
